[IMP] runbot: allow extra builds based on a cron
Actually, a build in a branch is created when a new commit is found.
When a user wants to schedule a time-based build, he has to generate a
commit outside of the runbot (e.g. in a cron script).

With this commit, cron builds can be scheduled hourly, daily, weekly or  monthly with extra parameters.